+++ tests/stats.rs
@@ -19,15 +19,10 @@ fn client_stats() {
 
     server_app.connect_client(&mut client_app);
 
-    client_app
-        .world_mut()
-        .spawn((TestComponent, Signature::from(0)));
     let server_entity = server_app
         .world_mut()
         .spawn((Replicated, TestComponent, Signature::from(0)))
         .id();
-
-    server_app.world_mut().spawn(Replicated).despawn();
 
     server_app.update();
     server_app.exchange_with_client(&mut client_app);
@@ -43,14 +38,21 @@ fn client_stats() {
     server_app.update();
     server_app.exchange_with_client(&mut client_app);
     client_app.update();
+    server_app.exchange_with_client(&mut client_app);
+
+    server_app.world_mut().despawn(server_entity);
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
 
     let stats = client_app.world().resource::<ClientReplicationStats>();
     assert_eq!(stats.entities_changed, 2);
     assert_eq!(stats.components_changed, 2);
     assert_eq!(stats.mappings, 1);
     assert_eq!(stats.despawns, 1);
-    assert_eq!(stats.messages, 2);
-    assert_eq!(stats.bytes, 24);
+    assert_eq!(stats.messages, 3);
+    assert_eq!(stats.bytes, 25);
 }
 
 #[derive(Component, Deserialize, Serialize)]