+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move `VisibilityPolicy` to `server::client_visibility` module.
 - Include `TrackAppExt::track_mutate_messages` in the replication protocol, since it affects the serialization format.
 
+### Fixed
+
+- Avoid sending despawn for entities that weren't sent.
+
 ## [0.36.1] - 2025-10-11
 
 ### Changed